+++ cmd/CHANGE_LOG.md
@@ -1,2 +1,10 @@
+------------------------------------------------------------------------------------------------------------------------
+Date: 2025.08.13
+Version: v0.1.0-alpha.2
+Changes:
+- add configuration for command: command description; params description, required, type, default value
+- add auto validation for params
+
+------------------------------------------------------------------------------------------------------------------------
 Date: 2025.07.24
 Version: v0.1.0-alpha.1